バグ等修正

+++ app/javascript/packs/post.js
@@ -14,38 +14,40 @@ const handleHeartDisplay = (hasLiked) => {
 document.addEventListener('turbolinks:load', () => {
 
     const dataset = $('#post-show').data()
-    const postId = dataset.postId
+    if (dataset) {
+        const postId = dataset.postId
 
-    axios.get(`/api/posts/${postId}/like`).then((response) => {
-        const hasLiked = response.data.hasLiked
-        handleHeartDisplay(hasLiked)
-    })
+        axios.get(`/api/posts/${postId}/like`).then((response) => {
+            const hasLiked = response.data.hasLiked
+            handleHeartDisplay(hasLiked)
+        })
 
-    $('.inactive-heart').on('click', () => {
-        axios.post(`/api/posts/${postId}/like`)
-        .then((response) => {
-            if (response.data.status == 'ok' ) {
-                $('.active-heart').removeClass('hidden')
-                $('.inactive-heart').addClass('hidden')
-            }
+        $('.inactive-heart').on('click', () => {
+            axios.post(`/api/posts/${postId}/like`)
+            .then((response) => {
+                if (response.data.status == 'ok' ) {
+                    $('.active-heart').removeClass('hidden')
+                    $('.inactive-heart').addClass('hidden')
+                }
+            })
+            .catch((e) => {
+                window.alert('Error')
+                console.log(e)
+            })
         })
-        .catch((e) => {
-            window.alert('Error')
-            console.log(e)
-        })
-    })
 
-    $('.active-heart').on('click', () => {
-        axios.delete(`/api/posts/${postId}/like`)
-        .then((response) => {
-            if (response.data.status == 'ok' ) {
-                $('.inactive-heart').removeClass('hidden')
-                $('.active-heart').addClass('hidden')
-            }
+        $('.active-heart').on('click', () => {
+            axios.delete(`/api/posts/${postId}/like`)
+            .then((response) => {
+                if (response.data.status == 'ok' ) {
+                    $('.inactive-heart').removeClass('hidden')
+                    $('.active-heart').addClass('hidden')
+                }
+            })
+            .catch((e) => {
+                window.alert('Error')
+                console.log(e)
+            })
         })
-        .catch((e) => {
-            window.alert('Error')
-            console.log(e)
-        })
-    })
+    }
 })
